fix size_t issue for clang

+++ src/RVmodel.h
@@ -134,11 +134,11 @@ class RVmodel
 
 
         // priors for KO mode!
-        std::vector<std::shared_ptr<DNest4::ContinuousDistribution>> KO_Pprior {n_known_object};
-        std::vector<std::shared_ptr<DNest4::ContinuousDistribution>> KO_Kprior {n_known_object};
-        std::vector<std::shared_ptr<DNest4::ContinuousDistribution>> KO_eprior {n_known_object};
-        std::vector<std::shared_ptr<DNest4::ContinuousDistribution>> KO_phiprior {n_known_object};
-        std::vector<std::shared_ptr<DNest4::ContinuousDistribution>> KO_wprior {n_known_object};
+        std::vector<std::shared_ptr<DNest4::ContinuousDistribution>> KO_Pprior {static_cast<size_t>(n_known_object)};
+        std::vector<std::shared_ptr<DNest4::ContinuousDistribution>> KO_Kprior {static_cast<size_t>(n_known_object)};
+        std::vector<std::shared_ptr<DNest4::ContinuousDistribution>> KO_eprior {static_cast<size_t>(n_known_object)};
+        std::vector<std::shared_ptr<DNest4::ContinuousDistribution>> KO_phiprior {static_cast<size_t>(n_known_object)};
+        std::vector<std::shared_ptr<DNest4::ContinuousDistribution>> KO_wprior {static_cast<size_t>(n_known_object)};
